Removing bogus circuit sketch, adding new functional sketch
This is the start of a new chapter in this project's development. I just tested the mechanism for the first time, and it works. The sketch has basic functionality EXCEPT that it has no interface to JMRI, and it has no calibration/memory system, so it's VERY easy to mess up the positioning, and the only way to fix that is to open up the mechanism and tune it back into place.