Reorganizing, half-started CMRI interface
Arduino IDE insisted I save the .ino file in a folder that matched, so I obliged (mostly to shut it up). I also started the very beginning of a CMRI interface, but did not have time to do any kind of testing or even finish a simple test sketch. The program is probably not going to compile correctly right now, or at the very least it won't run correctly.

Other than that, no changes since last commit.