Reliability fix, removed CMRI interface temporarily for debugging
I have removed the CMRI interface code temporarily as I implemented a reliability improvement. Sometimes, the hall effect sensor would trigger almost immediately when the motor started (since the magnet stopped almost directly above it), which isn't supposed to happen. So it would mess up the gear turn count. I added a time-based detection system for when that happens so that the Arduino notices when it's triggered unusually fast, and ignores it. I did a fairly lengthy test (about 10 minutes solid) and had no alignment issues at all, so I'm calling this fixed.