Committing half-working CMRI interface for use in an ArduinoCMRI issue
The low-level electromechanical part of the turntable is pretty much complete. There is rarely (if ever) an alignment error, and most of the ones that do come up are probably due to a buggy PC interface. I am making this commit so that I can access the sketch from GitHub and discuss it in issues.