Functional non-blocking rotation system and 180 degree rotate function
This commit is by no means a "polished" finished product, but it is a big step forward in development. Everything has been rewritten (aside from startup_beep() ) to be completely non-blocking, thus fixing the errors in the JMRI console. Also, in tinkering with the "rotate 180 degrees" function, I found out that for some reason turn_cw() and turn_ccw() dislike the number 12. If I simply set the targetGearTurns to 60 (12*5), the program works. I'm not calling that a permanent solution, but it is a step in the right direction and it works for now. I plan on polishing up the user-editable configuration variables at the top, and just generally making everything more user-friendly over the next few commits.